#kupac prikaz titule kupca, obracunavanje popusta i totalne cene sa popustom, izracunavanje bodova potrebnih da se dostigne nova titula

+++ veb/WebShopREST 2/src/beans/BuyerType.java
@@ -5,7 +5,8 @@ import enums.BuyerTitle;
 public class BuyerType {
     private BuyerTitle title;
     private int discount;
-    private int points;
+    private double points;
+    
 	public BuyerTitle getTitle() {
 		return title;
 	}
@@ -18,13 +19,13 @@ public class BuyerType {
 	public void setDiscount(int discount) {
 		this.discount = discount;
 	}
-	public int getPoints() {
+	public double getPoints() {
 		return points;
 	}
-	public void setPoints(int points) {
+	public void setPoints(double points) {
 		this.points = points;
 	}
-	public BuyerType(BuyerTitle title, int discount, int points) {
+	public BuyerType(BuyerTitle title, int discount, double points) {
 		super();
 		this.title = title;
 		this.discount = discount;
